requirements.txt
azure-storage-blob==12.20.0
geopandas==0.13.2
google_api_python_client==2.98.0
google_auth_oauthlib==1.1.0
numpy==1.25.1
ocha_anticipy==1.1.3
pandas==2.0.3
protobuf==4.24.4
python-dotenv==1.0.0
python_dateutil==2.8.2
Requests==2.31.0
Shapely==2.0.2
tqdm==4.66.1
